+++ infra/__main__.py
@@ -11,46 +11,46 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 import os
-import pathlib
 import sys
 
 import datarobot as dr
 import pulumi
 import pulumi_datarobot as datarobot
+from datarobot_pulumi_utils.common.feature_flags import check_feature_flags
+from datarobot_pulumi_utils.common.urls import get_deployment_url
+from datarobot_pulumi_utils.pulumi.custom_model_deployment import CustomModelDeployment
+from datarobot_pulumi_utils.pulumi.proxy_llm_blueprint import ProxyLLMBlueprint
+from datarobot_pulumi_utils.pulumi.stack import PROJECT_NAME
+from datarobot_pulumi_utils.schema.apps import CustomAppResourceBundles
+from datarobot_pulumi_utils.schema.llms import LLMs
 
 sys.path.append("..")
+
+from settings_main import PROJECT_ROOT
+
 from infra import (
     settings_app_infra,
     settings_generative,
 )
-from infra.common.feature_flags import check_feature_flags
-from infra.common.globals import GlobalLLM, GlobalRuntimeEnvironment
-from infra.common.stack import PROJECT_ROOT, project_name
-from infra.common.urls import get_deployment_url
-from infra.components.custom_model_deployment import CustomModelDeployment
 from infra.components.dr_credential import (
     get_credential_runtime_parameter_values,
     get_database_credentials,
     get_llm_credentials,
 )
-from infra.components.proxy_llm_blueprint import ProxyLLMBlueprint
 from infra.settings_database import DATABASE_CONNECTION_TYPE
 from infra.settings_proxy_llm import CHAT_MODEL_NAME
 from utils.resources import (
     app_env_name,
     llm_deployment_env_name,
 )
-from utils.schema import (
-    AppInfra,
-)
+from utils.schema import AppInfra
 
 TEXTGEN_DEPLOYMENT_ID = os.environ.get("TEXTGEN_DEPLOYMENT_ID")
 TEXTGEN_REGISTERED_MODEL_ID = os.environ.get("TEXTGEN_REGISTERED_MODEL_ID")
 
 
-if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM:
+if settings_generative.LLM == LLMs.DEPLOYED_LLM:
     pulumi.info(f"{TEXTGEN_DEPLOYMENT_ID=}")
     pulumi.info(f"{TEXTGEN_REGISTERED_MODEL_ID=}")
     if (TEXTGEN_DEPLOYMENT_ID is None) == (TEXTGEN_REGISTERED_MODEL_ID is None):  # XOR
@@ -58,9 +58,7 @@ if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM:
             "Either TEXTGEN_DEPLOYMENT_ID or TEXTGEN_REGISTERED_MODEL_ID must be set when using a deployed LLM. Plese check your .env file"
         )
 
-check_feature_flags(
-    pathlib.Path(PROJECT_ROOT / "infra" / "feature_flag_requirements.yaml")
-)
+check_feature_flags(PROJECT_ROOT / "infra" / "feature_flag_requirements.yaml")
 
 with open(PROJECT_ROOT / "frontend/app_infra.json", "w") as infra_selection:
     infra_selection.write(
@@ -79,12 +77,12 @@ if "DATAROBOT_DEFAULT_USE_CASE" in os.environ:
     )
 else:
     use_case = datarobot.UseCase(
-        resource_name=f"Data Analyst Use Case [{project_name}]",
+        resource_name=f"Data Analyst Use Case [{PROJECT_NAME}]",
         description="Use case for Data Analyst application",
     )
 
 prediction_environment = datarobot.PredictionEnvironment(
-    resource_name=f"Data Analyst Prediction Environment [{project_name}]",
+    resource_name=f"Data Analyst Prediction Environment [{PROJECT_NAME}]",
     platform=dr.enums.PredictionEnvironmentPlatform.DATAROBOT_SERVERLESS,
 )
 
@@ -99,7 +97,7 @@ playground = datarobot.Playground(
     **settings_generative.playground_args.model_dump(),
 )
 
-if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM:
+if settings_generative.LLM == LLMs.DEPLOYED_LLM:
     if TEXTGEN_REGISTERED_MODEL_ID is not None:
         proxy_llm_registered_model = datarobot.RegisteredModel.get(
             resource_name="Existing TextGen Registered Model",
@@ -107,10 +105,10 @@ if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM:
         )
 
         proxy_llm_deployment = datarobot.Deployment(
-            resource_name=f"Data Analyst LLM Deployment [{project_name}]",
+            resource_name=f"Data Analyst LLM Deployment [{PROJECT_NAME}]",
             registered_model_version_id=proxy_llm_registered_model.version_id,
             prediction_environment_id=prediction_environment.id,
-            label=f"Data Analyst LLM Deployment [{project_name}]",
+            label=f"Data Analyst LLM Deployment [{PROJECT_NAME}]",
             use_case_ids=[use_case.id],
             opts=pulumi.ResourceOptions(
                 replace_on_changes=["registered_model_version_id"]
@@ -132,7 +130,7 @@ if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM:
         **settings_generative.llm_blueprint_args.model_dump(mode="python"),
     )
 
-elif settings_generative.LLM != GlobalLLM.DEPLOYED_LLM:
+elif settings_generative.LLM != LLMs.DEPLOYED_LLM:
     llm_blueprint = datarobot.LlmBlueprint(  # type: ignore[assignment]
         playground_id=playground.id,
         **settings_generative.llm_blueprint_args.model_dump(),
@@ -143,12 +141,12 @@ llm_custom_model = datarobot.CustomModel(
     use_case_ids=[use_case.id],
     source_llm_blueprint_id=llm_blueprint.id,
     runtime_parameter_values=[]
-    if settings_generative.LLM == GlobalLLM.DEPLOYED_LLM
+    if settings_generative.LLM == LLMs.DEPLOYED_LLM
     else llm_runtime_parameter_values,
 )
 
 llm_deployment = CustomModelDeployment(
-    resource_name=f"Chat Agent Deployment [{project_name}]",
+    resource_name=f"Chat Agent Deployment [{PROJECT_NAME}]",
     use_case_ids=[use_case.id],
     custom_model_version_id=llm_custom_model.version_id,
     registered_model_args=settings_generative.registered_model_args,
@@ -178,9 +176,8 @@ app_source = datarobot.ApplicationSource(
         runtime_parameter_values=app_runtime_parameters
     ),
     runtime_parameter_values=app_runtime_parameters,
-    base_environment_id=GlobalRuntimeEnvironment.PYTHON_312_APPLICATION_BASE.value.id,
     resources=datarobot.ApplicationSourceResourcesArgs(
-        resource_label="cpu.xlarge",
+        resource_label=CustomAppResourceBundles.CPU_XL.value.id,
     ),
     **settings_app_infra.app_source_args,
 )