+++ infra/settings_proxy_llm.py
@@ -17,34 +17,33 @@ import os
 
 import datarobot as dr
 import pulumi
-
-from infra.common.globals import GlobalRuntimeEnvironment
-from infra.common.stack import project_name
-from utils.schema import LLMDeploymentSettings
-
-from .common.schema import (
+from datarobot_pulumi_utils.pulumi.stack import PROJECT_NAME
+from datarobot_pulumi_utils.schema.custom_models import (
     CustomModelArgs,
     DeploymentArgs,
     RegisteredModelArgs,
 )
+from datarobot_pulumi_utils.schema.exec_envs import RuntimeEnvironments
+
+from utils.schema import LLMDeploymentSettings
 
 CHAT_MODEL_NAME = os.getenv("CHAT_MODEL_NAME")
 
 custom_model_args = CustomModelArgs(
-    resource_name=f"Data Analyst Proxy LLM Custom Model [{project_name}]",
-    name=f"Data Analyst Proxy LLM Custom Model [{project_name}]",
+    resource_name=f"Data Analyst Proxy LLM Custom Model [{PROJECT_NAME}]",
+    name=f"Data Analyst Proxy LLM Custom Model [{PROJECT_NAME}]",
     target_name=LLMDeploymentSettings().target_feature_name,
     target_type=dr.enums.TARGET_TYPE.TEXT_GENERATION,
     replicas=2,
-    base_environment_id=GlobalRuntimeEnvironment.PYTHON_312_MODERATIONS.value.id,
+    base_environment_id=RuntimeEnvironments.PYTHON_312_MODERATIONS.value.id,
     opts=pulumi.ResourceOptions(delete_before_replace=True),
 )
 
 registered_model_args = RegisteredModelArgs(
-    resource_name=f"Data Analyst Proxy LLM Registered Model [{project_name}]",
+    resource_name=f"Data Analyst Proxy LLM Registered Model [{PROJECT_NAME}]",
 )
 
 deployment_args = DeploymentArgs(
-    resource_name=f"Data Analyst Proxy LLM Deployment [{project_name}]",
-    label=f"Data Analyst Proxy LLM Deployment [{project_name}]",
+    resource_name=f"Data Analyst Proxy LLM Deployment [{PROJECT_NAME}]",
+    label=f"Data Analyst Proxy LLM Deployment [{PROJECT_NAME}]",
 )