+++ utils/schema.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 from __future__ import annotations
 
 import json
@@ -51,7 +50,7 @@ class LLMDeploymentSettings(BaseModel):
     prompt_feature_name: str = "promptText"
 
 
-class AiCatalogDataset(BaseModel):
+class DataRegistryDataset(BaseModel):
     id: str
     name: str
     created: str